small fixes

+++ util/single_use/zinc.py
@@ -5,6 +5,7 @@ import re
 import tempfile
 from pathlib import Path
 from tqdm import tqdm
+import csv
 
 import pandas as pd
 
@@ -26,12 +27,16 @@ def main(wget_fname, output_label, wdir='wdir'):
 
     with open(wget_fname, 'r') as f: commands = f.read().splitlines()
     with open(log_fname, 'r') as f: logged_labels = f.read().splitlines()
-
     staged_labels = []
     staged_info = []
 
+    with open(smi_output_fname, 'w') as f:
+        writer = csv.writer(f, delimiter=' ')
+        writer.writerow(['smiles', 'zinc_id'])
+
     logger.info(f"wget file: {wget_fname}")
-    for command in commands:
+    for idx, command in enumerate(commands):
+        logger.info(f'{command}, {idx}')
         label = get_label(command)
         if is_logged(label, logged_labels, staged_labels):
             continue
@@ -40,14 +45,38 @@ def main(wget_fname, output_label, wdir='wdir'):
                                         wdir=wdir))
         staged_labels.append(label)
 
+        if idx%10 == 9:
+            write_entries(staged_info, smi_output_fname, staged_labels,
+                          log_fname)
+            staged_info, staged_labels, logged_labels = reset_staged(
+                log_fname)
+
+
+    write_entries(staged_info, smi_output_fname, staged_labels,
+                          log_fname)
+
+def reset_staged(log_fname):
+    with open(log_fname, 'r') as f:
+        logged_labels = f.read().splitlines()
+    return [], [], logged_labels
+
+
+def write_entries(staged_info, smi_output_fname, staged_labels, log_fname):
+
+    staged_info = [i for i in staged_info if i is not None]
     if len(staged_info) != 0 :
         staged_info_df = pd.concat(staged_info, ignore_index=True)
-        staged_info_df.to_csv(smi_output_fname, mode='a', sep=' ')
+        staged_info_df.to_csv(smi_output_fname, mode='a', sep=' ',
+                              index=False, header=False)
 
-        with open(log_fname, 'a') as f:
-            f.write('\n'.join(staged_labels)+ '\n')
+        logger.info(f"saved {len(staged_info)} entries")
     else:
         logger.info("no new entries found")
+
+    with open(log_fname, 'a') as f:
+        f.write('\n'.join(staged_labels) + '\n')
+
+
 
 def only_has_CH(entry):
     return not bool(re.search(r'[a-bd-gi-zA-BD-GI-Z]', entry))
@@ -68,13 +97,21 @@ def collect_info(command, label, wget_stdout, wget_stderr, wdir):
         f.write(result.stdout)
     with open(wget_stderr, 'a') as f:
         f.write(result.stderr)
-    
-    data = pd.read_csv(tmp_fname, delim_whitespace=True)
 
-    data = filter_for_CH(data)
+    filesize = Path(tmp_fname).stat().st_size
+    if filesize > 0:
+        data = pd.read_csv(tmp_fname, delim_whitespace=True)
 
-    os.remove(tmp_fname)
-    return data
+        data = filter_for_CH(data)
+
+        if len(data) == 0:
+            data = None
+
+        os.remove(tmp_fname)
+        return data
+    else:
+        logger.info(f"Found no entries in {command}")
+        return None
 
 
 def is_logged(label, logged_labels, staged_labels):
